Reuse vmstructs implementation from upstream
