+++ ddprof-lib/src/main/cpp/javaApi.cpp
@@ -25,7 +25,7 @@
 #include "thread.h"
 #include "tsc.h"
 #include "vmEntry.h"
-#include "vmStructs.h"
+#include "vmStructs_dd.h"
 #include "wallClock.h"
 #include <errno.h>
 #include <fstream>
@@ -312,7 +312,7 @@ Java_com_datadoghq_profiler_JVMAccess_findStringJVMFlag0(JNIEnv *env,
                                                          jobject unused,
                                                          jstring flagName) {
   JniString flag_str(env, flagName);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(), {JVMFlag::Type::String});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(), {ddprof::JVMFlag::Type::String});
   if (f) {
     char** value = static_cast<char**>(f->addr());
     if (value != NULL && *value != NULL) {
@@ -329,7 +329,7 @@ Java_com_datadoghq_profiler_JVMAccess_setStringJVMFlag0(JNIEnv *env,
                                                          jstring flagValue) {
   JniString flag_str(env, flagName);
   JniString value_str(env, flagValue);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(), {JVMFlag::Type::String});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(), {ddprof::JVMFlag::Type::String});
   if (f) {
     char** value = static_cast<char**>(f->addr());
     if (value != NULL) {
@@ -343,7 +343,7 @@ Java_com_datadoghq_profiler_JVMAccess_findBooleanJVMFlag0(JNIEnv *env,
                                                          jobject unused,
                                                          jstring flagName) {
   JniString flag_str(env, flagName);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(), {JVMFlag::Type::Bool});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(), {ddprof::JVMFlag::Type::Bool});
   if (f) {
     char* value = static_cast<char*>(f->addr());
     if (value != NULL) {
@@ -359,7 +359,7 @@ Java_com_datadoghq_profiler_JVMAccess_setBooleanJVMFlag0(JNIEnv *env,
                                                          jstring flagName,
                                                          jboolean flagValue) {
   JniString flag_str(env, flagName);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(), {JVMFlag::Type::Bool});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(), {ddprof::JVMFlag::Type::Bool});
   if (f) {
     char* value = static_cast<char*>(f->addr());
     if (value != NULL) {
@@ -373,7 +373,7 @@ Java_com_datadoghq_profiler_JVMAccess_findIntJVMFlag0(JNIEnv *env,
                                                          jobject unused,
                                                          jstring flagName) {
   JniString flag_str(env, flagName);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(), {JVMFlag::Type::Int, JVMFlag::Type::Uint, JVMFlag::Type::Intx, JVMFlag::Type::Uintx, JVMFlag::Type::Uint64_t, JVMFlag::Type::Size_t});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(), {ddprof::JVMFlag::Type::Int, ddprof::JVMFlag::Type::Uint, ddprof::JVMFlag::Type::Intx, ddprof::JVMFlag::Type::Uintx, ddprof::JVMFlag::Type::Uint64_t, ddprof::JVMFlag::Type::Size_t});
   if (f) {
     long* value = static_cast<long*>(f->addr());
     if (value != NULL) {
@@ -388,7 +388,7 @@ Java_com_datadoghq_profiler_JVMAccess_findFloatJVMFlag0(JNIEnv *env,
                                                          jobject unused,
                                                          jstring flagName) {
   JniString flag_str(env, flagName);
-  JVMFlag *f = JVMFlag::find(flag_str.c_str(),{ JVMFlag::Type::Double});
+  ddprof::JVMFlag *f = ddprof::JVMFlag::find(flag_str.c_str(),{ ddprof::JVMFlag::Type::Double});
   if (f) {
     double* value = static_cast<double*>(f->addr());
     if (value != NULL) {
